ibc_app_nft_transfer
Module types

Re-exports the implementation of the IBC Non-Fungible Token Transfer (ICS-721) data structures.

Modules

error

Defines the Non-Fungible Token Transfer (ICS-721) error types.

events

Defines Non-Fungible Token Transfer (ICS-721) event types.

msgs

Defines the Non-Fungible Token Transfer (ICS-721) message types.

packet

Contains the PacketData type that defines the structure of NFT transfers’ packet bytes

proto

Re-exports ICS-721 NFT transfer proto types from the ibc-proto crate.

Structs

ClassData

Class data for an NFT

ClassId

Class ID for an NFT

ClassUri

Class URI for an NFT

Data

DataValue

Ics721Data

Memo

Represents the token transfer memo

PrefixedClassId

Prefixed class to trace sources like ICS-20 PrefixedDenom

TokenData

Token data for an NFT

TokenId

Token ID for an NFT

TokenIds

TokenUri

Token URI for an NFT

TracePath

A full trace path modelled as a collection of TracePrefixs.

TracePrefix

One hop in a token’s trace, which consists of the port and channel IDs of the sender

Constants

ACK_SUCCESS_B64

The successful string, used for creating an acknowledgement status. It is equivalent to base64::encode(0x01).

MODULE_ID_STR

Module identifier for the ICS-721 application.

PORT_ID_STR

The port identifier that the ICS-721 applications typically bind with.

VERSION

ICS-721 application current version.

Functions

ack_success_b64

Returns a successful acknowledgement status for the NFT transfer application.

is_receiver_chain_source

Returns true if the class ID originally came from the receiving chain, and false otherwise.

is_sender_chain_source

Returns true if the class ID originally came from the sender chain, and false otherwise.